#8 - Improve comment highlighting when selected
